# Revert "BasisTranslator to cache bound target DAG. (#6187)"